+++ document/python_package/conf.py
@@ -85,7 +85,13 @@ todo_include_todos = False
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = 'default'
+on_rtd = os.environ.get("READTHEDOCS") == "True"
+
+if on_rtd:
+    html_theme = 'default'
+else:
+    html_theme = 'alabaster'
+
 
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
@@ -166,7 +172,8 @@ def run_apidoc(_):
     for package_name in package_names:
         output_pathname = os.path.join(cwd, "api", package_name)
 
-        main(["--separate", "-o", output_pathname, os.path.join(source_pathname, package_name), "--force"])
+        main(["--separate", "-o", output_pathname,
+            os.path.join(source_pathname, package_name), "--force"])
 
 
 def setup(app):
@@ -174,5 +181,6 @@ def setup(app):
 
 
 intersphinx_mapping = {
+    "flask": ("http://flask.pocoo.org/docs/0.12", None),
     "python": ("https://docs.python.org/3.5", None)
 }