+++ document/python_package/changes.rst
@@ -2,8 +2,14 @@ Changes
 =======
 
 
+0.0.2
+-----
+- Add ``geoserver`` sub-package
+- Add ``flask.error_handler`` sub-package
+
+
 0.0.1
-----------
+-----
 - Add ``format_time_point`` template filter
 - Add ``utc_now`` function